backend/requirements.txt
fastapi[standard]
uvicorn
openai
python-dotenv
